requirements.txt
streamlit==1.33.0
pandas==2.2.2
numpy==1.26.4

langchain==1.2.15
langchain-core==1.2.28
langchain-community==0.4.1
langchain-groq==1.1.2
langgraph==1.1.6

sentence-transformers==2.7.0
faiss-cpu==1.8.0

tavily-python==0.7.23
python-dotenv==1.2.2
requests==2.31.0